whisprr-backend/requirements.txt
fastapi[all]
google-generativeai 
python-dotenv
sqlmodel